Component Composition - Props Between Tags
<TabButton>Some Text</TabButton>
To do this you have to tell you component what to do with that text. Whats inside is always assigned to the children property of props.
It can also be a JSX structure.
Use it instead of attributes is a matter of preference. Maybe you want them to be similar to some other HTML component.