State is implemented to change the topic selected.
State
You cannot simply change a variable in a component and expect React to know. The component is only rendered once. To make it redere again you have to tell React to do it. Thats where State comes in. This controls a variable from react letting it know when it has to render again. To do this, you use the useState hook.
You set a default value inside the function.
- useState('default');
It always return an array with 2 values so is common tu Destructure it.
- The first one is the current value.
- The second one is a function to be used to update the state, and re-render. This executes the component function again with the new value.
- When you call the function, the value is only updated after the component function is called again.
```
const [selecterTopic, setSelectedTopic ] = useState('Please Click an Option');
```