Implemented Conditional Rednering
You can implement conditional rendering with ternary expressions. Just one liners.
The results in any JSX have to be in one parent node.
```
          {!selectedTopic ? <p>Please Select a Topic</p> :
            <div id="tab-content">
              <h3>{EXAMPLES[selectedTopic].title}</h3>
              <p>{EXAMPLES[selectedTopic].description}</p>
              <pre>
                <code>
                  {EXAMPLES[selectedTopic].code}
                </code>
              </pre>
            </div>
          }
```
You can also use the lazy operator. Since the first part is false, the second one is not evaluated, is not executed.
{!selectedTopic} && <p>Please Select a topic</p>
It has to be 1 liner, or, you can also use variables with JSX code inside.