Added Conditional Fromatting for the Selected Button
Conditional Formatting
in JSX you have tu use className for CSS class.
We use a variable sent as a prop to the component to tell which style. In the exaple we use a boolean.
In the component call we add the prop for the selected based on if the button is selected.