Extracted CoreConcepts and Examples to their own components.
In the orignal App component, whenever you change an Example tab, second line after the title changed because all components where re-rendered, and the first word depends on a random number. By extracting them, we now have, the state only in examples, and the core concepts are not re-rendered when the example is changed.