Prop Fowarding
Props are not fowarded automatically to children, you have to receive them and use them manually
But you can use Fowarder Props. This is done by using the spread operator when destructuring, and then again in the styles.

We implemented this for the id and class on Examples, and the function that handles the click in TabButton.